refactor(zona): reorganizar excepciones y ajustar servicio de zona
