+++ src/main/java/com/playko/zoologico/exception/zona/ZonaNotFoundException.java
@@ -1,4 +1,4 @@
-package com.playko.zoologico.exception;
+package com.playko.zoologico.exception.zona;
 
 public class ZonaNotFoundException extends RuntimeException{
     public ZonaNotFoundException() {
